+++ accueil/models/__init__.py
@@ -1,0 +1,14 @@
+from .autre import Creneau,Programme,Ramassage,JourFerie,Message,Destinataire,MatiereECTS,NoteECTS,mois
+from .config import Config
+from .matiere import Matiere
+from .classe import Classe
+from .groupe import Groupe
+from .etablissement import Etablissement
+from .colleur import Colleur 
+from .user import User
+from .prof import Prof
+from .semaine import Semaine
+from .note import Note
+from .colle import Colle
+from .eleve import Eleve
+
